+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project are documented in this file.
 
 
+## 2025-12-08
+### Added
+- `client.js` to allow host swapping
+
+
 ## 2025-11-24
 ### Removed
 - Dev docker files